+++ dbfacts.py
@@ -247,9 +247,6 @@ def main ():
     # This will be the name you reference in Ansible. i.e. source_facts['sga_target'] (source_facts)
     refname = ""
 
-    # os.system("/usr/bin/scl enable python27 bash")
-    # os.system("scl enable python27 bash")
-
     module = AnsibleModule(
       argument_spec = dict(
         systempwd       =dict(required=True),