+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [v2.0.0] - 2021-11-15
+### Module fixes associated with v2.0.0 and the massive merge of cru-ansible-oracle
+- mkalias fixes to work on disaster recovery environment
+- linkmapper fixes
+- removed all references to os.system("/usr/bin/scl enable python27 bash") to correct way: #!/usr/bin/env python3
+
+## [v2.0.0] - 2021-11-15
 ### Created new modules escpass.py, linkmapper.py, xtractitem.py
 - escpass.py looks for special characters that need escaped to be used by oracle in a password.
 - linkmapper.py takes a python list of filters and database link mappings and remaps current db links and synonyms to reflect mapping changes.
@@ -82,9 +88,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   more dynamic than Ansible. The functionality was moved to utils GUI app
   which resides in the DBA Google drive.
 
-[Unreleased]: https://github.com/CruGlobal/cru-ansible-modules/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/CruGlobal/cru-ansible-modules/compare/v2.0.1...HEAD
 
-[v1.5.0]: https://github.com/CruGlobal/cru-ansible-modules/compare/v1.4.0...v2.0.0
+[v2.0.1]: https://github.com/CruGlobal/cru-ansible-modules/compare/v2.0.0...v2.0.1
+[v2.0.0]: https://github.com/CruGlobal/cru-ansible-modules/compare/v1.4.0...v2.0.0
 [v1.4.0]: https://github.com/CruGlobal/cru-ansible-modules/compare/v1.3.5...v1.4.0
 [v1.3.5]: https://github.com/CruGlobal/cru-ansible-modules/compare/v1.3.4...v1.3.5
 [v1.3.4]: https://github.com/CruGlobal/cru-ansible-modules/compare/v1.3.3...v1.3.4